repair delete_avatars bug.

+++ WeShare/models.py
@@ -277,6 +277,7 @@ class Notification(db.Model):
 def delete_avatars(**kwargs):
     target = kwargs['target']
     for filename in [target.avatar_s, target.avatar_m, target.avatar_l, target.avatar_raw]:
-        path = os.path.join(current_app.config['AVATARS_SAVE_PATH'], filename)
-        if os.path.exists(path):
-            os.remove(path)
+        if filename:
+            path = os.path.join(current_app.config['AVATARS_SAVE_PATH'], filename)
+            if os.path.exists(path):
+                os.remove(path)
